requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
python-dotenv